+++ packages/mobile/src/screens/ConnectionProcess/ConnectionProcess.screen.tsx
@@ -1,5 +1,5 @@
 import React, { FC, useCallback, useEffect } from 'react'
-import { communities, connection, ErrorCodes, errors, publicChannels, users } from '@quiet/state-manager'
+import { connection, ErrorCodes, errors } from '@quiet/state-manager'
 import { useDispatch, useSelector } from 'react-redux'
 import ConnectionProcessComponent from '../../components/ConnectionProcess/ConnectionProcess.component'
 import { Linking } from 'react-native'
@@ -9,17 +9,10 @@ import { ScreenNames } from '../../const/ScreenNames.enum'
 export const ConnectionProcessScreen: FC = () => {
   const dispatch = useDispatch()
 
-  const connectionProcessSelector = useSelector(connection.selectors.torConnectionProcess)
+  const connectionProcessSelector = useSelector(connection.selectors.connectionProcess)
+  const isJoiningCompletedSelector = useSelector(connection.selectors.isJoiningCompleted)
+
   const error = useSelector(errors.selectors.registrarErrors)
-
-  const community = useSelector(communities.selectors.currentCommunity)
-  const isOwner = Boolean(community?.CA)
-
-  const channelsStatusSorted = useSelector(publicChannels.selectors.channelsStatusSorted)
-  const messageNotNull = channelsStatusSorted.filter(channel => channel.newestMessage !== undefined)
-
-  const certificatesMapping = useSelector(users.selectors.certificatesMapping)
-  const channels = useSelector(publicChannels.selectors.publicChannels)
 
   const openUrl = useCallback((url: string) => {
     void Linking.openURL(url)
@@ -36,17 +29,14 @@ export const ConnectionProcessScreen: FC = () => {
   }, [error, dispatch])
 
   useEffect(() => {
-    const areChannelsLoaded = channels.length > 0
-    const areCertificatesLoaded = Object.values(certificatesMapping).length > 0
-    const isAllDataLoaded = areChannelsLoaded && areCertificatesLoaded
-    console.log({ areChannelsLoaded, areCertificatesLoaded })
-    if (isOwner ? connectionProcessSelector.number == 85 : isAllDataLoaded && messageNotNull.length !== 0) {
+    if (isJoiningCompletedSelector) {
       dispatch(
         navigationActions.replaceScreen({
           screen: ScreenNames.ChannelListScreen,
         })
       )
     }
-  }, [connectionProcessSelector, messageNotNull, certificatesMapping, channels])
+  }, [isJoiningCompletedSelector])
+
   return <ConnectionProcessComponent openUrl={openUrl} connectionProcess={connectionProcessSelector} />
 }