+++ packages/state-manager/src/sagas/network/network.selectors.ts
@@ -15,9 +15,7 @@ export const initializedCommunities = createSelector(networkSlice, reducerState 
 export const isCurrentCommunityInitialized = createSelector(
   initializedCommunities,
   currentCommunity,
-  (initializedCommunities, currentCommunity) => {
-    return currentCommunity && initializedCommunities[currentCommunity.id]
-  }
+  (initializedCommunities, currentCommunity) => currentCommunity && initializedCommunities[currentCommunity.id]
 )
 
 export const connectedPeers = createSelector(networkSlice, reducerState => {